moves registry factory to mod
